Seed + SyncDB + Subplatformdependant notifications fix

+++ BL/ItemManager.cs
@@ -3,7 +3,9 @@ using PB.BL.Domain.Dashboards;
 using PB.BL.Domain.Items;
 using PB.BL.Domain.JSONConversion;
 using PB.BL.Domain.Platform;
+using PB.BL.Domain.Settings;
 using PB.BL.Interfaces;
+using PB.BL.RestClient;
 using PB.DAL;
 using PB.DAL.Interfaces;
 using System;
@@ -30,7 +32,6 @@ namespace PB.BL
             UowManager = uowMgr;
             ItemRepo = new ItemRepo(uowMgr.UnitOfWork);
             RecordRepo = new RecordRepo(uowMgr.UnitOfWork);
-
         }
 
         #region Init
@@ -359,24 +360,86 @@ namespace PB.BL
         }
         #endregion
 
-        public void Seed(bool evenRecords = true)
+        #region Keywords
+        public IEnumerable<Keyword> GetKeywords()
+        {
+            return ItemRepo.ReadKeywords();
+        }
+        #endregion
+
+        public void SyncDatabase(Subplatform subplatform)
         {
             InitNonExistingRepo();
-            List<Record> toegevoegde = JClassToRecord(RecordRepo.Seed(evenRecords));
+
+            // Validation
+            if (subplatform.Settings.FirstOrDefault(ss => ss.SettingName == Setting.Platform.DAYS_TO_KEEP_RECORDS) is null)
+            {
+                throw new Exception("Subplatform has no set period to keep records!");
+            }
+            if (subplatform.Settings.FirstOrDefault(ss => ss.SettingName == Setting.Platform.SOURCE_API_URL) is null)
+            {
+                throw new Exception("Subplatform has no set API url!");
+            }
+
+            // Injects api seed data
+            APICalls restClient = new APICalls()
+            {
+                API_URL = subplatform.Settings.FirstOrDefault(ss => ss.SettingName == Setting.Platform.SOURCE_API_URL).Value
+            };
+
+            // Call API with request
+            List<JClass> requestedRecords = new List<JClass>();
+            try
+            {
+                requestedRecords.AddRange(restClient.RequestRecords(since: DateTime.Now.AddDays(-int.Parse(subplatform.Settings.First(s => s.SettingName.Equals(Setting.Platform.DAYS_TO_KEEP_RECORDS)).Value))));
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.GetType().Name + ": " + e.Message);
+                if (e.InnerException != null) Console.WriteLine("Inner Exception: " + e.InnerException);
+                throw (e);
+            }
+
+            // Link items to subplatform
+            requestedRecords.ForEach(r => r.Subplatforms.Add(subplatform));
+
+            //Convert JClass to Record and persist to database
+            List<Record> newRecords = JClassToRecord(requestedRecords);
+
+            // Persist items tgo database
+            RecordRepo.CreateRecords(newRecords);
+
+            // Save pending changes
+            UowManager.Save();
         }
 
-        public void CleanupOldRecords(Subplatform subplatform, int days)
+        public void CleanupOldRecords(Subplatform subplatform)
         {
             InitNonExistingRepo();
-            List<Person> persons = ItemRepo.ReadPersons().Where(i => i.SubPlatforms.Contains(subplatform)).ToList();
-            List<Record> oldRecords = new List<Record>();
 
-            persons.ForEach(p =>
+            // Validation
+            if (subplatform.Settings.FirstOrDefault(ss => ss.SettingName == Setting.Platform.DAYS_TO_KEEP_RECORDS) is null)
             {
-                oldRecords.AddRange(p.Records.Where(r => r.Date.Date < DateTime.Now.Date.AddDays(-days)));
-            });
+                throw new Exception("Subplatform has no set period to keep records!");
+            }
+            
+            // Days to keep records
+            int days = int.Parse(subplatform.Settings.FirstOrDefault(ss => ss.SettingName == Setting.Platform.DAYS_TO_KEEP_RECORDS).Value);
 
+            // All persons
+            List<Person> persons = ItemRepo.ReadPersons().Where(i => i.SubPlatforms.Contains(subplatform)).ToList();
+
+            // All records to clear
+            List<Record> oldRecords = persons
+                .SelectMany(p => p.Records)
+                .Distinct()
+                .Where(r => r.Date.Date < DateTime.Now.Date.AddDays(-days))
+                .ToList();
+
+            // Persist deleted records
             RecordRepo.DeleteRecords(oldRecords);
+
+            // Save pending changes
             UowManager.Save();
         }
 
@@ -524,18 +587,9 @@ namespace PB.BL
                 }
             }
 
-            RecordRepo.CreateRecords(newRecords);
-            UowManager.Save();
-
             return newRecords;
         }
 
-        #region Keywords
-        public IEnumerable<Keyword> GetKeywords()
-        {
-            return ItemRepo.ReadKeywords();
-        }
-        #endregion
 
         public int GetKeywordsCount()
         {
